backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
python-docx
PyPDF2
chardet
